+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @creezio/auth
 
+## 0.18.0
+
+### Patch Changes
+
+- Updated dependencies [7c40c12]
+  - @creezio/shell-ui@0.18.0
+  - @creezio/shell@0.18.0
+  - @creezio/platform-core@0.18.0
+
 ## 0.17.1
 
 ### Patch Changes