+++ packages/database/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @creezio/database
 
+## 0.18.0
+
+### Patch Changes
+
+- @creezio/platform-core@0.18.0
+
 ## 0.17.1
 
 ### Patch Changes